backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
cryptography==45.0.6
distro==1.9.0
dnspython==2.6.1
email-validator==2.3.0
exceptiongroup==1.3.0
faiss-cpu==1.8.0
fastapi==0.111.0
fastapi-cli==0.0.8
filelock==3.16.1
frozenlist==1.5.0
fsspec==2025.3.0
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
jiter==0.9.1
joblib==1.4.2
lxml==6.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.1
numpy==1.22.2
openai==0.28.0
orjson==3.10.15
packaging==25.0
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.4.0
propcache==0.2.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
pypdfium2==4.30.0
pytesseract==0.3.10
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.1.0
rich-toolkit==0.15.0
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.0.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
typer==0.16.1
typing_extensions==4.13.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.30.0
watchfiles==0.24.0
websockets==13.1
yarl==1.15.2
